requirements.txt
aiogram==3.0.0b7
requests==2.31.0
python-dotenv==1.0.0
Pillow==10.3.0
numpy==1.23.2